+++ src/lib/wrappers/AVUtil/AVFrameWrapper.cpp
@@ -20,15 +20,15 @@ namespace
 
 struct AVFrame_54
 {
-  uint8_t      *data[AV_NUM_DATA_POINTERS];
+  uint8_t *     data[AV_NUM_DATA_POINTERS];
   int           linesize[AV_NUM_DATA_POINTERS];
-  uint8_t     **extended_data;
+  uint8_t **    extended_data;
   int           width, height;
   int           nb_samples;
   int           format;
   int           key_frame;
   AVPictureType pict_type;
-  uint8_t      *base[AV_NUM_DATA_POINTERS];
+  uint8_t *     base[AV_NUM_DATA_POINTERS];
   AVRational    sample_aspect_ratio;
   int64_t       pts;
   int64_t       pkt_pts;
@@ -41,9 +41,9 @@ struct AVFrame_54
 
 struct AVFrame_55
 {
-  uint8_t      *data[AV_NUM_DATA_POINTERS];
+  uint8_t *     data[AV_NUM_DATA_POINTERS];
   int           linesize[AV_NUM_DATA_POINTERS];
-  uint8_t     **extended_data;
+  uint8_t **    extended_data;
   int           width, height;
   int           nb_samples;
   int           format;
@@ -63,9 +63,9 @@ typedef AVFrame_55 AVFrame_56;
 
 struct AVFrame_57
 {
-  uint8_t                           *data[AV_NUM_DATA_POINTERS];
+  uint8_t *                          data[AV_NUM_DATA_POINTERS];
   int                                linesize[AV_NUM_DATA_POINTERS];
-  uint8_t                          **extended_data;
+  uint8_t **                         extended_data;
   int                                width, height;
   int                                nb_samples;
   int                                format;
@@ -78,7 +78,7 @@ struct AVFrame_57
   int                                coded_picture_number;
   int                                display_picture_number;
   int                                quality;
-  void                              *opaque;
+  void *                             opaque;
   int                                repeat_pict;
   int                                interlaced_frame;
   int                                top_field_first;
@@ -86,10 +86,10 @@ struct AVFrame_57
   int64_t                            reordered_opaque;
   int                                sample_rate;
   uint64_t                           channel_layout;
-  AVBufferRef                       *buf[AV_NUM_DATA_POINTERS];
-  AVBufferRef                      **extended_buf;
+  AVBufferRef *                      buf[AV_NUM_DATA_POINTERS];
+  AVBufferRef **                     extended_buf;
   int                                nb_extended_buf;
-  AVFrameSideData                  **side_data;
+  AVFrameSideData **                 side_data;
   int                                nb_side_data;
   int                                flags;
   enum AVColorRange                  color_range;
@@ -100,7 +100,7 @@ struct AVFrame_57
   int64_t                            best_effort_timestamp;
   int64_t                            pkt_pos;
   int64_t                            pkt_duration;
-  AVDictionary                      *metadata;
+  AVDictionary *                     metadata;
 
   // Actually, there is more here, but the variables above are the only we need.
 };
@@ -109,7 +109,7 @@ typedef AVFrame_57 AVFrame_58;
 
 } // namespace
 
-AVFrameWrapper::AVFrameWrapper(AVFrame                                  *frame,
+AVFrameWrapper::AVFrameWrapper(AVFrame *                                 frame,
                                std::shared_ptr<FFmpegLibrariesInterface> librariesInterface)
     : frame(frame), librariesInterface(librariesInterface)
 {
@@ -149,7 +149,7 @@ Size AVFrameWrapper::getSize() const
 
 int64_t AVFrameWrapper::getPTS() const
 {
-  int pts;
+  int64_t pts;
   CAST_AVUTIL_GET_MEMBER(AVFrame, this->frame, pts, pts);
   return pts;
 }
